Update maven artifacts to same version (#108)
using this plugin with maven 3.9 or better results in warnings due to using invalid alignment of maven in the plugin.  This is regardless of maven artifacts being provided.  This update brings all the artifacts for maven core up to date at 3.9.11 and updates the tests and dependency lock files for same.